requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
api==0.0.7
appdirs==1.4.3            # via black
appnope==0.1.0            # via ipython
attrs==19.1.0             # via black
backcall==0.1.0           # via ipython
black==19.3b0
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
click==7.0                # via black
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
decorator==4.3.2          # via ipython, traitlets
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django==2.1.7
djangorestframework==3.9.1
entrypoints==0.3          # via flake8
flake8==3.7.7
idna==2.8                 # via requests
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==7.3.0            # via ipdb
itypes==1.1.0             # via coreapi
jedi==0.13.3              # via ipython
jinja2==2.10.1            # via coreschema
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.1    # via mypy
mypy==0.720
nose==1.3.7               # via api
openapi-codec==1.3.2      # via django-rest-swagger
parso==0.3.4              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
prompt-toolkit==2.0.9     # via ipython
psycopg2==2.7.7
ptyprocess==0.6.0         # via pexpect
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.0           # via flake8
pygments==2.3.1           # via ipython
pytz==2018.9              # via django
requests==2.21.0
simplejson==3.16.0        # via django-rest-swagger
six==1.12.0               # via django-rest-auth, prompt-toolkit, traitlets
toml==0.10.0              # via black
traitlets==4.3.2          # via ipython
typed-ast==1.4.0          # via mypy
typing-extensions==3.7.4  # via mypy
uritemplate==3.0.0        # via coreapi
urllib3==1.24.1           # via requests
wcwidth==0.1.7            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.1.0        # via ipdb, ipython
